+++ R/mo_property.R
@@ -27,12 +27,12 @@
 #'
 #' Use these functions to return a specific property of a microorganism based on the latest accepted taxonomy. All input values will be evaluated internally with [as.mo()], which makes it possible to use microbial abbreviations, codes and names as input. Please see *Examples*.
 #' @inheritSection lifecycle Stable lifecycle
-#' @param x any character (vector) that can be coerced to a valid microorganism code with [as.mo()]. Can be omitted for auto-guessing in `mo_is_*()` functions when used inside `dplyr` verbs, such as [`filter()`][dplyr::filter()], [`mutate()`][dplyr::mutate()] and [`summarise()`][dplyr::summarise()], please see *Examples*.
+#' @param x any character (vector) that can be coerced to a valid microorganism code with [as.mo()]. Can be omitted for auto-guessing the column containing microorganism codes when used inside `dplyr` verbs, such as [`filter()`][dplyr::filter()], [`mutate()`][dplyr::mutate()] and [`summarise()`][dplyr::summarise()], please see *Examples*.
 #' @param property one of the column names of the [microorganisms] data set: `r paste0('"``', colnames(microorganisms), '\``"', collapse = ", ")`, or must be `"shortname"`
 #' @param language language of the returned text, defaults to system language (see [get_locale()]) and can be overwritten by setting the option `AMR_locale`, e.g. `options(AMR_locale = "de")`, see [translate]. Also used to translate text like "no growth". Use `language = NULL` or `language = ""` to prevent translation.
 #' @param ... other arguments passed on to [as.mo()], such as 'allow_uncertain' and 'ignore_pattern'
 #' @param ab any (vector of) text that can be coerced to a valid antibiotic code with [as.ab()]
-#' @param open browse the URL using [utils::browseURL()]
+#' @param open browse the URL using [`browseURL()`][utils::browseURL()]
 #' @details All functions will return the most recently known taxonomic property according to the Catalogue of Life, except for [mo_ref()], [mo_authors()] and [mo_year()]. Please refer to this example, knowing that *Escherichia blattae* was renamed to *Shimwellia blattae* in 2010:
 #' - `mo_name("Escherichia blattae")` will return `"Shimwellia blattae"` (with a message about the renaming)
 #' - `mo_ref("Escherichia blattae")` will return `"Burgess et al., 1973"` (with a message about the renaming)
@@ -161,9 +161,13 @@
 #' mo_info("E. coli")
 #' }
 mo_name <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   translate_AMR(mo_validate(x = x, property = "fullname", language = language, ...), language = language, only_unknown = FALSE)
 }
 
@@ -174,22 +178,26 @@ mo_fullname <- mo_name
 #' @rdname mo_property
 #' @export
 mo_shortname <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   x.mo <- as.mo(x, language = language, ...)
-
+  
   metadata <- get_mo_failures_uncertainties_renamed()
-
+  
   replace_empty <- function(x) {
     x[x == ""] <- "spp."
     x
   }
-
+  
   # get first char of genus and complete species in English
   genera <- mo_genus(x.mo, language = NULL)
   shortnames <- paste0(substr(genera, 1, 1), ". ", replace_empty(mo_species(x.mo, language = NULL)))
-
+  
   # exceptions for where no species is known
   shortnames[shortnames %like% ".[.] spp[.]"] <- genera[shortnames %like% ".[.] spp[.]"]
   # exceptions for staphylococci
@@ -199,7 +207,8 @@ mo_shortname <- function(x, language = get_locale(), ...) {
   shortnames[shortnames %like% "S. group [ABCDFGHK]"] <- paste0("G", gsub("S. group ([ABCDFGHK])", "\\1", shortnames[shortnames %like% "S. group [ABCDFGHK]"]), "S")
   # unknown species etc.
   shortnames[shortnames %like% "unknown"] <- paste0("(", trimws(gsub("[^a-zA-Z -]", "", shortnames[shortnames %like% "unknown"])), ")")
-
+  
+  shortnames[is.na(x.mo)] <- NA_character_
   load_mo_failures_uncertainties_renamed(metadata)
   translate_AMR(shortnames, language = language, only_unknown = FALSE)
 }
@@ -207,72 +216,104 @@ mo_shortname <- function(x, language = get_locale(), ...) {
 #' @rdname mo_property
 #' @export
 mo_subspecies <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   translate_AMR(mo_validate(x = x, property = "subspecies", language = language, ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_species <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   translate_AMR(mo_validate(x = x, property = "species", language = language, ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_genus <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   translate_AMR(mo_validate(x = x, property = "genus", language = language, ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_family <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   translate_AMR(mo_validate(x = x, property = "family", language = language, ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_order <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   translate_AMR(mo_validate(x = x, property = "order", language = language, ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_class <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   translate_AMR(mo_validate(x = x, property = "class", language = language, ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_phylum <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   translate_AMR(mo_validate(x = x, property = "phylum", language = language, ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_kingdom <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   translate_AMR(mo_validate(x = x, property = "kingdom", language = language, ...), language = language, only_unknown = TRUE)
 }
 
@@ -283,21 +324,29 @@ mo_domain <- mo_kingdom
 #' @rdname mo_property
 #' @export
 mo_type <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   translate_AMR(mo_validate(x = x, property = "kingdom", language = language, ...), language = language, only_unknown = FALSE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_gramstain <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   x.mo <- as.mo(x, language = language, ...)
   metadata <- get_mo_failures_uncertainties_renamed()
-
+  
   x.phylum <- mo_phylum(x.mo)
   # DETERMINE GRAM STAIN FOR BACTERIA
   # Source: https://itis.gov/servlet/SingleRpt/SingleRpt?search_topic=TSN&search_value=956097
@@ -318,7 +367,7 @@ mo_gramstain <- function(x, language = get_locale(), ...) {
                     "Firmicutes",
                     "Tenericutes")
     | x.mo == "B_GRAMP"] <- "Gram-positive"
-
+  
   load_mo_failures_uncertainties_renamed(metadata)
   translate_AMR(x, language = language, only_unknown = FALSE)
 }
@@ -327,12 +376,12 @@ mo_gramstain <- function(x, language = get_locale(), ...) {
 #' @export
 mo_is_gram_negative <- function(x, language = get_locale(), ...) {
   if (missing(x)) {
-    # this supports using in dplyr verbs: ... %>% filter(mo_is_gram_negative())
-    x <- find_mo_col("mo_is_gram_negative")
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
   }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   x.mo <- as.mo(x, language = language, ...)
   metadata <- get_mo_failures_uncertainties_renamed()
   grams <- mo_gramstain(x.mo, language = NULL)
@@ -346,12 +395,12 @@ mo_is_gram_negative <- function(x, language = get_locale(), ...) {
 #' @export
 mo_is_gram_positive <- function(x, language = get_locale(), ...) {
   if (missing(x)) {
-    # this supports using in dplyr verbs: ... %>% filter(mo_is_gram_positive())
-    x <- find_mo_col("mo_is_gram_positive")
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
   }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   x.mo <- as.mo(x, language = language, ...)
   metadata <- get_mo_failures_uncertainties_renamed()
   grams <- mo_gramstain(x.mo, language = NULL)
@@ -399,27 +448,39 @@ mo_is_intrinsic_resistant <- function(x, ab, language = get_locale(), ...) {
 #' @rdname mo_property
 #' @export
 mo_snomed <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   mo_validate(x = x, property = "snomed", language = language, ...)
 }
 
 #' @rdname mo_property
 #' @export
 mo_ref <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   mo_validate(x = x, property = "ref", language = language, ...)
 }
 
 #' @rdname mo_property
 #' @export
 mo_authors <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   x <- mo_validate(x = x, property = "ref", language = language, ...)
   # remove last 4 digits and presumably the comma and space that preceed them
   x[!is.na(x)] <- gsub(",? ?[0-9]{4}", "", x[!is.na(x)])
@@ -429,9 +490,13 @@ mo_authors <- function(x, language = get_locale(), ...) {
 #' @rdname mo_property
 #' @export
 mo_year <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   x <- mo_validate(x = x, property = "ref", language = language, ...)
   # get last 4 digits
   x[!is.na(x)] <- gsub(".*([0-9]{4})$", "\\1", x[!is.na(x)])
@@ -441,21 +506,29 @@ mo_year <- function(x, language = get_locale(), ...) {
 #' @rdname mo_property
 #' @export
 mo_rank <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   mo_validate(x = x, property = "rank", language = language, ...)
 }
 
 #' @rdname mo_property
 #' @export
 mo_taxonomy <- function(x, language = get_locale(),  ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   x <- as.mo(x, language = language, ...)
   metadata <- get_mo_failures_uncertainties_renamed()
-
+  
   result <- list(kingdom = mo_kingdom(x, language = language),
                  phylum = mo_phylum(x, language = language),
                  class = mo_class(x, language = language),
@@ -464,7 +537,7 @@ mo_taxonomy <- function(x, language = get_locale(),  ...) {
                  genus = mo_genus(x, language = language),
                  species = mo_species(x, language = language),
                  subspecies = mo_subspecies(x, language = language))
-
+  
   load_mo_failures_uncertainties_renamed(metadata)
   result
 }
@@ -472,12 +545,16 @@ mo_taxonomy <- function(x, language = get_locale(),  ...) {
 #' @rdname mo_property
 #' @export
 mo_synonyms <- function(x, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   x <- as.mo(x, language = language, ...)
   metadata <- get_mo_failures_uncertainties_renamed()
-
+  
   IDs <- mo_name(x = x, language = NULL)
   syns <- lapply(IDs, function(newname) {
     res <- sort(microorganisms.old[which(microorganisms.old$fullname_new == newname), "fullname"])
@@ -493,7 +570,7 @@ mo_synonyms <- function(x, language = get_locale(), ...) {
   } else {
     result <- unlist(syns)
   }
-
+  
   load_mo_failures_uncertainties_renamed(metadata)
   result
 }
@@ -501,12 +578,16 @@ mo_synonyms <- function(x, language = get_locale(), ...) {
 #' @rdname mo_property
 #' @export
 mo_info <- function(x, language = get_locale(),  ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   x <- as.mo(x, language = language, ...)
   metadata <- get_mo_failures_uncertainties_renamed()
-
+  
   info <- lapply(x, function(y)
     c(mo_taxonomy(y, language = language),
       list(synonyms = mo_synonyms(y),
@@ -519,7 +600,7 @@ mo_info <- function(x, language = get_locale(),  ...) {
   } else {
     result <- info[[1L]]
   }
-
+  
   load_mo_failures_uncertainties_renamed(metadata)
   result
 }
@@ -527,14 +608,18 @@ mo_info <- function(x, language = get_locale(),  ...) {
 #' @rdname mo_property
 #' @export
 mo_url <- function(x, open = FALSE, language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(open, allow_class = "logical", has_length = 1)
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   mo <- as.mo(x = x, language = language, ... = ...)
   mo_names <- mo_name(mo)
   metadata <- get_mo_failures_uncertainties_renamed()
-
+  
   df <- data.frame(mo, stringsAsFactors = FALSE) %pm>%
     pm_left_join(pm_select(microorganisms, mo, source, species_id), by = "mo")
   df$url <- ifelse(df$source == "CoL",
@@ -544,14 +629,14 @@ mo_url <- function(x, open = FALSE, language = get_locale(), ...) {
                           NA_character_))
   u <- df$url
   names(u) <- mo_names
-
+  
   if (open == TRUE) {
     if (length(u) > 1) {
       warning_("Only the first URL will be opened, as `browseURL()` only suports one string.")
     }
     utils::browseURL(u[1L])
   }
-
+  
   load_mo_failures_uncertainties_renamed(metadata)
   u
 }
@@ -560,21 +645,25 @@ mo_url <- function(x, open = FALSE, language = get_locale(), ...) {
 #' @rdname mo_property
 #' @export
 mo_property <- function(x, property = "fullname", language = get_locale(), ...) {
+  if (missing(x)) {
+    # this supports using in dplyr verbs: ... %>% filter(mo_is_intrinsic_resistant(ab = "amox"))
+    x <- find_mo_col("mo_is_intrinsic_resistant")
+  }
   meet_criteria(x, allow_NA = TRUE)
   meet_criteria(property, allow_class = "character", has_length = 1, is_in = colnames(microorganisms))
   meet_criteria(language, has_length = 1, is_in = c(LANGUAGES_SUPPORTED, ""), allow_NULL = TRUE, allow_NA = TRUE)
-
+  
   translate_AMR(mo_validate(x = x, property = property, language = language, ...), language = language, only_unknown = TRUE)
 }
 
 mo_validate <- function(x, property, language, ...) {
   check_dataset_integrity()
-
+  
   if (tryCatch(all(x[!is.na(x)] %in% MO_lookup$mo) & length(list(...)) == 0, error = function(e) FALSE)) {
     # special case for mo_* functions where class is already <mo>
     return(MO_lookup[match(x, MO_lookup$mo), property, drop = TRUE])
   }
-
+  
   dots <- list(...)
   Becker <- dots$Becker
   if (is.null(Becker)) {
@@ -584,12 +673,12 @@ mo_validate <- function(x, property, language, ...) {
   if (is.null(Lancefield)) {
     Lancefield <- FALSE
   }
-
+  
   # try to catch an error when inputting an invalid argument
   # so the 'call.' can be set to FALSE
   tryCatch(x[1L] %in% MO_lookup[1, property, drop = TRUE],
            error = function(e) stop(e$message, call. = FALSE))
-
+  
   if (is.mo(x)
       & !Becker %in% c(TRUE, "all")
       & !Lancefield %in% c(TRUE, "all")) {
@@ -601,7 +690,7 @@ mo_validate <- function(x, property, language, ...) {
              | Lancefield %in% c(TRUE, "all")) {
     x <- exec_as.mo(x, property = property, language = language, ...)
   }
-
+  
   if (property == "mo") {
     return(set_clean_class(x, new_class = c("mo", "character")))
   } else if (property == "snomed") {
@@ -614,13 +703,16 @@ mo_validate <- function(x, property, language, ...) {
 find_mo_col <- function(fn) {
   # this function tries to find an mo column using dplyr::cur_data_all() for mo_is_*() functions,
   # which is useful when functions are used within dplyr verbs
-  df <- get_current_data("x", call = -3) # will return an error if not found
+  df <- get_current_data(arg_name = "x", call = -3) # will return an error if not found
   mo <- NULL
   try({
     mo <- suppressMessages(search_type_in_df(df, "mo"))
   }, silent = TRUE)
   if (!is.null(df) && !is.null(mo) && is.data.frame(df)) {
-    message_("Using column '", font_bold(mo), "' as input for ", fn, "()")
+    if (message_not_thrown_before(fn = fn)) {
+      message_("Using column '", font_bold(mo), "' as input for ", fn, "()")
+      remember_thrown_message(fn = fn)
+    }
     return(df[, mo, drop = TRUE])
   } else {
     stop_("argument `x` is missing and no column with info about microorganisms could be found.", call = -2)